content-security-policy/frame-src/frame-src-5_1.html

<!DOCTYPE HTML>
<html>
<head>
    <title>iframe element src attribute must match src list</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <link rel='stylesheet' href='/resources/testharness.css'>
</head>
<body>
<h1>iframe element src attribute must match src list</h1>
<div id='log'></div>

<script>
var selfTest = async_test("Iframe loaded from self (as allowed by frame-src policy)");
var nonSelfTest = async_test("Iframe not loaded from www1.web-platform.test (as denied by frame-src policy)");
setTimeout(function () {nonSelfTest.done();}, 2000);
var selfTestAgain = async_test("second valid iframe test");
setTimeout(function () {selfTestAgain.done();}, 2000);
</script>

<iframe src="frame-src-5_1yes.html" 
    id="selfiFrame" 
    onload='selfTest.done()'>
</iframe>
<iframe src="http://www1.web-platform.test:8000/content-security-policy/frame-src/frame-src-5_1no.html"
    id="nonSelfiFrame"
    onload='nonSelfTest.step(function() {assert_false(true, "loaded bad iframe from www1.web-platform.test (should have been denied by policy)")})'>
</iframe>
<iframe src="http://web-platform.test:8000/content-security-policy/frame-src/frame-src-5_1yes.html"
    id="selfiFrame2"
    onload='selfTestAgain.step(function() {assert_true(true, "selfiFrame loaded")})'>
</iframe>

<script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=frame-src%20%27self%27'></script>
</body>
</html>
